add checks before user deletion

+++ register/views.py
@@ -16,7 +16,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.messages.views import SuccessMessageMixin
 from django.http import Http404
-
+from reservation.models import ReservationPost
 
 def register(request):
     # Redirect to login page
@@ -162,15 +162,36 @@ class DonorUpdateView(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
 
 def delete_profile(request):
     user = request.user
-    try:
+    if user.donorprofile:
+    	reservationposts = ReservationPost.objects.filter(donor=user)
+    elif user.helpseekerprofile:
+    	reservationposts = ReservationPost.objects.filter(helpseeker=user)
+    confirmed = 0
+
+    for reservationpost in reservationposts:
+        print(reservationpost.post.status)
+        if reservationpost.post.status == "RESERVED":
+            confirmed += 1
+    if confirmed == 0:
+        for reservationpost in reservationposts:
+            if reservationpost.post.status != "CLOSED" and reservationpost.post.status != "RESERVED":
+                reservationpost.post.status = "AVAILABLE"
+                reservationpost.post.save()
+                print(reservationpost.post.status)
         user.delete()
         messages.success(request, "Account deleted successfully.")
         return redirect("/")
-    except Exception:
+    elif confirmed > 0:
         messages.error(
-            request, "Your profile deletion was unsuccessful. Please try again!"
-        )
+        request, "You can not delete your profile because you have " + str(confirmed) + " confirmed reservation" + "s." if confirmed > 0 else "."
+    )
         return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+    else:
+        messages.error(
+        request, "Your profile deletion was unsuccessful. Please try again!"
+    )
+        return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+
 
 
 def bad_request(request, exception):
